Refactor: useWindowSize로 리팩토링
