+++ src/components/WindowSizeExample.jsx
@@ -1,27 +1,8 @@
-import { useEffect, useState } from "react";
+import {useWindowSize} from "../hooks/useWindowSize"
 
 export const WindowSizeExample = () => {
   // 실습 1. 하단 코드를 useWindowSize (커스텀 훅으로 바꿔주세요!)
-  const [windowSize, setWindowSize] = useState({
-    width: window.innerWidth,
-    height: window.innerHeight,
-  });
-
-  useEffect(() => {
-    const handleResize = () => {
-      setWindowSize({
-        width: window.innerWidth,
-        height: window.innerHeight,
-      });
-    };
-
-    window.addEventListener("resize", handleResize);
-
-    return () => {
-      window.removeEventListener("resize", handleResize);
-    };
-  }, []);
-
+  const {windowSize}=useWindowSize();
   return (
     <div>
       <h2>useWindowSize 실습</h2>
